Hook up Receive step to a database for persistence
This makes the app feel like it's actually doing something;
we can see it importing sample data from a stubbed-out system.

Right now the server is merely connecting queries to the built-in Rails
postgres database. To make the system more general-purpose, we'll want
to change it up to route every request to an appropriate Docker
container.

For now, however, everything works well enough.