Switch back to `localstorage.removeItem()`
It turns out `localStorage.item = null` doesn't have the same effect
as `localStorage.removeItem()`.

With `localStorage.item = null`, we were getting
"Signed in as null" displaying in the sidebar,
along with the link to sign out.
There was no way to clear out the signed-in "null" user,
except for manually deleting the localStorage variables
through the developer tools.

The other direct calls to `localStorage` seem to work fine.

Bug introduced in #54.