On each of the tab pages, expand the linked sample
E.g. if you visit `/test/Q123456`
and the test page includes has sample `Q123456`,
the sample will be expanded by default.

To make this work, the `Sample` component
looks at the closest matching route's parameters.

If it has a parameter called `subpath` equal to the sample's ID,
then the sample will render expanded by default.

This is a bit messy, as it relies on a keyword.
However, it might turn out to be generic enough
to be useful in a variety of situations.

If it ain't broke, don't fix it.